Fetch integer as i64
SQLite stores integers in 64 bit.

If we try to get i32, it is just [cast internally](https://docs.rs/libsql/0.9.2/src/libsql/rows.rs.html#169-177) as i32.

The logic for fallible int conversion is already written. This commit
makes use of it and makes it explicit that we are actually retrieving an i64 from the database before converting it to our model type.

+++ crates/devolutions-pedm/src/db/libsql.rs
@@ -41,7 +41,7 @@ impl Database for LibsqlConn {
         let version = self
             .query_one("SELECT version FROM pedm_schema_version", ())
             .await?
-            .get::<i32>(0)?;
+            .get::<i64>(0)?;
         Ok(i16::try_from(version)?)
     }
